# Extend iterateAvailablePlans to handle events.