--- a/src/core/algorithm/plan.cc
+++ b/src/core/algorithm/plan.cc
@@ -1,5 +1,6 @@
 #include "core/algorithm/plan.h"
 
+#include <glog/logging.h>
 #include <iostream>
 #include <sstream>
 
@@ -70,9 +71,11 @@ std::string RefPlan::decisionText() const
 }
 
 template<typename Callback>
-void iterateAvailablePlansInternal(const CoreField& field,
+void iterateAvailablePlansInternal(CoreField field,
                                    const KumipuyoSeq& kumipuyoSeq,
                                    std::vector<Decision>& decisions,
+                                   const std::vector<Plan::Event>& events,
+                                   size_t eventIndex,
                                    int currentDepth,
                                    int maxDepth,
                                    int currentNumChigiri,
@@ -90,6 +93,16 @@ void iterateAvailablePlansInternal(const CoreField& field,
     } else {
         ptr = ALL_KUMIPUYO_KINDS;
         n = 10;
+    }
+
+    if (eventIndex < events.size()) {
+        const Plan::Event& event = events[eventIndex];
+        if (event.type == Plan::Event::Type::FALL_OJAMA_ROWS && event.frames < totalFrames) {
+            totalFrames += field.fallOjama(event.value);
+            ++eventIndex;
+        }
+        if (!field.isEmpty(3, 12))
+            return;
     }
 
     for (int j = 0; j < 22; j++) {
@@ -118,8 +131,8 @@ void iterateAvailablePlansInternal(const CoreField& field,
             if (currentDepth + 1 == maxDepth || shouldFire) {
                 callback(nextField, decisions, currentNumChigiri + isChigiri, totalFrames, dropFrames, shouldFire);
             } else {
-                iterateAvailablePlansInternal(nextField, kumipuyoSeq, decisions, currentDepth + 1, maxDepth,
-                                              currentNumChigiri + isChigiri, totalFrames + dropFrames, callback);
+                iterateAvailablePlansInternal(nextField, kumipuyoSeq, decisions, events, eventIndex, currentDepth + 1,
+                                              maxDepth, currentNumChigiri + isChigiri, totalFrames + dropFrames, callback);
             }
         }
         decisions.pop_back();
@@ -131,6 +144,17 @@ void Plan::iterateAvailablePlans(const CoreField& field,
                                  const KumipuyoSeq& kumipuyoSeq,
                                  int maxDepth,
                                  const Plan::IterationCallback& callback)
+{
+    std::vector<Event> events;
+    iterateAvailablePlansWithEvents(field, kumipuyoSeq, maxDepth, events, callback);
+}
+
+// static
+void Plan::iterateAvailablePlansWithEvents(const CoreField& field,
+                                           const KumipuyoSeq& kumipuyoSeq,
+                                           int maxDepth,
+                                           const std::vector<Event>& events,
+                                           const Plan::IterationCallback& callback)
 {
     std::vector<Decision> decisions;
     decisions.reserve(maxDepth);
@@ -155,7 +179,7 @@ void Plan::iterateAvailablePlans(const CoreField& field,
         }
     };
 
-    iterateAvailablePlansInternal(field, kumipuyoSeq, decisions, 0, maxDepth, 0, 0, f);
+    iterateAvailablePlansInternal(field, kumipuyoSeq, decisions, events, 0, 0, maxDepth, 0, 0, f);
 }
 
 // static
@@ -164,7 +188,18 @@ void Plan::iterateAvailablePlansWithoutFiring(const CoreField& field,
                                               int maxDepth,
                                               const Plan::RensaIterationCallback& callback)
 {
+    std::vector<Event> events;
+    iterateAvailablePlansWithoutFiringWithEvents(field, kumipuyoSeq, maxDepth, events, callback);
+}
+
+// static
+void Plan::iterateAvailablePlansWithoutFiringWithEvents(const CoreField& field,
+                                                        const KumipuyoSeq& kumipuyoSeq,
+                                                        int maxDepth,
+                                                        const std::vector<Event>& events,
+                                                        const Plan::RensaIterationCallback& callback)
+{
     std::vector<Decision> decisions;
     decisions.reserve(maxDepth);
-    iterateAvailablePlansInternal(field, kumipuyoSeq, decisions, 0, maxDepth, 0, 0, callback);
+    iterateAvailablePlansInternal(field, kumipuyoSeq, decisions, events, 0, 0, maxDepth, 0, 0, callback);
 }