--- a/src/core/algorithm/plan.h
+++ b/src/core/algorithm/plan.h
@@ -15,6 +15,22 @@ class RefPlan;
 
 class Plan {
 public:
+    // Event which may happen during the iteration.
+    // Extend Type if you want to use more types of events.
+    struct Event {
+        enum Type {
+            FALL_OJAMA_ROWS,
+        };
+
+        // This figures when this event will happen. 0 figures the frameId when a API is called.
+        int frames;
+
+        // Meaning of this variable depends on which type of event.
+        int value;
+
+        Type type;
+    };
+
     Plan() {}
     Plan(const CoreField& field, const std::vector<Decision>& decisions,
          const RensaResult& rensaResult, int numChigiri, int framesToIgnite, int lastDropFrames,
@@ -29,10 +45,14 @@ public:
     typedef std::function<void (const RefPlan&)> IterationCallback;
     // if |kumipuyos.size()| < |depth|, we will add extra kumipuyo.
     static void iterateAvailablePlans(const CoreField&, const KumipuyoSeq&, int depth, const IterationCallback&);
+    static void iterateAvailablePlansWithEvents(const CoreField&, const KumipuyoSeq&, int depth,
+                                                const std::vector<Event>& events, const IterationCallback&);
 
     typedef std::function<void (const CoreField&, const std::vector<Decision>&,
                                 int numChigiri, int framesToIgnite, int lastDropFrames, bool shouldFire)> RensaIterationCallback;
     static void iterateAvailablePlansWithoutFiring(const CoreField&, const KumipuyoSeq&, int depth, const RensaIterationCallback&);
+    static void iterateAvailablePlansWithoutFiringWithEvents(const CoreField&, const KumipuyoSeq&, int depth,
+                                                             const std::vector<Event>& events, const RensaIterationCallback&);
 
     const CoreField& field() const { return field_; }