--- a/src/core/algorithm/plan_test.cc
+++ b/src/core/algorithm/plan_test.cc
@@ -70,14 +70,45 @@ TEST(Plan, iterateAvailablePlansWithEvents)
 
     // With |events|, 2 rows of Ojama will fall just after 1st control,
     // so we cannot vanish puyos, because of ojama wall.
-    std::vector<Plan::Event> events = {Plan::Event {1, 2, Plan::Event::Type::FALL_OJAMA_ROWS}};
+    std::vector<Plan::Event> events = {Plan::Event::fallOjamaRowsEvent(1, 2)};
     found = false;
     Plan::iterateAvailablePlansWithEvents(field, seq, 2, events, callback);
     EXPECT_FALSE(found);
 
     // With |lateEvents|, Ojama will fall 1000 frames later.  It means we can control
     // 2 Kumipuyos before Ojama fall, and we can vanish puyos.
-    std::vector<Plan::Event> lateEvents = {Plan::Event {1000, 2, Plan::Event::Type::FALL_OJAMA_ROWS}};
+    std::vector<Plan::Event> lateEvents = {Plan::Event::fallOjamaRowsEvent(1000, 2)};
+    found = false;
+    Plan::iterateAvailablePlansWithEvents(field, seq, 2, lateEvents, callback);
+    EXPECT_TRUE(found);
+}
+
+TEST(Plan, iterateAvailablePlansWithMultipleEvents)
+{
+    CoreField field;
+    KumipuyoSeq seq("BBBB");
+    bool found = false;
+    
+    auto callback = [&found](const RefPlan& p) {
+        if (p.isRensaPlan())
+            found = true;
+    };
+
+    // Two rows of ojama in total will fall before 2nd control.
+    std::vector<Plan::Event> events = {
+        Plan::Event::fallOjamaRowsEvent(1, 1),
+        Plan::Event::fallOjamaRowsEvent(10, 1)
+    };
+    found = false;
+    Plan::iterateAvailablePlansWithEvents(field, seq, 2, events, callback);
+    EXPECT_FALSE(found);
+
+    // A row of ojama fall before 2nd control, and another row of
+    // ojama will fall after 2nd control.
+    std::vector<Plan::Event> lateEvents = {
+        Plan::Event::fallOjamaRowsEvent(1, 1),
+        Plan::Event::fallOjamaRowsEvent(1000, 1)
+    };
     found = false;
     Plan::iterateAvailablePlansWithEvents(field, seq, 2, lateEvents, callback);
     EXPECT_TRUE(found);