Using the new APIs, we will be able to fall rows of ojama puyos during the iterations.
Performance regression seems small enough to be ignored.

Performance tests:
before
```
[ RUN      ] PlanPerformanceTest.Filled23
        N = 10
      min = 3359484
      max = 5519563
  average = 3.85792e+06
deviation = 645352
[       OK ] PlanPerformanceTest.Filled23 (14 ms)
[ RUN      ] PlanPerformanceTest.Empty24
        N = 10
      min = 362338851
      max = 373247613
  average = 3.66979e+08
deviation = 3.17033e+06
[       OK ] PlanPerformanceTest.Empty24 (1362 ms)
[ RUN      ] PlanPerformanceTest.Filled24
        N = 10
      min = 428158605
      max = 468231909
  average = 4.35227e+08
deviation = 1.13371e+07
[       OK ] PlanPerformanceTest.Filled24 (1615 ms)
[----------] 6 tests from PlanPerformanceTest (3138 ms total)
```

after
```
[ RUN      ] PlanPerformanceTest.Filled23
        N = 10
      min = 3349941
      max = 4855311
  average = 3.77028e+06
deviation = 570955
[       OK ] PlanPerformanceTest.Filled23 (14 ms)
[ RUN      ] PlanPerformanceTest.Empty24
        N = 10
      min = 362581797
      max = 414601665
  average = 3.7237e+08
deviation = 1.66012e+07
[       OK ] PlanPerformanceTest.Empty24 (1382 ms)
[ RUN      ] PlanPerformanceTest.Filled24
        N = 10
      min = 427249824
      max = 438439887
  average = 4.31608e+08
deviation = 3.39203e+06
[       OK ] PlanPerformanceTest.Filled24 (1603 ms)
[----------] 6 tests from PlanPerformanceTest (3148 ms total)
```